⚡️ Improve media subtype on iOS (#863)

+++ ios/Classes/core/PMAssetPathEntity.h
@@ -32,7 +32,7 @@
 @property(nonatomic, assign) double lat;
 @property(nonatomic, assign) double lng;
 @property(nonatomic, copy) NSString *title;
-@property(nonatomic, assign) int subtype;
+@property(nonatomic, assign) NSUInteger subtype;
 @property(nonatomic, assign) BOOL favorite;
 @property(nonatomic, assign) BOOL isLocallyAvailable;
